more frontend awesomeness

+++ src/main/scala/omnibus/domain/topic/TopicStates.scala
@@ -2,7 +2,10 @@ package omnibus.domain.topic
 
 import omnibus.domain._
 
-case class TopicStatisticState(topic:String, throughputPerSec : Long, subscribersNumber: Long, subTopicsNumber : Long,
+case class TopicStatisticState(topic:String
+	                         , throughputPerSec : Double
+	                         , subscribersNumber: Long
+	                         , subTopicsNumber : Long,
                                timestamp: Long = System.currentTimeMillis / 1000)
 
 case class TopicState(events: List[Message] = Nil) {
